requirements.txt
pydantic==2.11.7
pytest==8.4.1
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
matplotlib==3.8.4
pyyaml==6.0.2
tqdm==4.67.1

# === CPU-only ===
torch==2.2.2
torchvision==0.17.2

# # # === GPU with CUDA 12.1 ===
# # # NOTE: If you use this option, install with:
# # #     pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu121
# torch==2.2.2+cu121
# torchvision==0.17.2+cu121
